Display dialog for local storage not available
If local storage is not available for some reason the user will be
displayed a dialog informing them that if they refresh or reload their
browser all data will be deleted.

Update validate list and todo, ensuring lists or todos not get created
when the user closes the validation alert.